etl/requirements.txt
pandas==2.0.1
mysql-connector-python==8.0.33
SQLAlchemy==2.0.15
psycopg2-binary==2.9.6